# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Urbano Express Orders API",
        "description": "API REST para gestión de órdenes de e-commerce",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8888",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/api/orders": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Listar todas las órdenes",
                "description": "Obtiene una lista de todas las órdenes con sus items asociados",
                "operationId": "955bb77a45ae173a485fadb8befed93f",
                "responses": {
                    "200": {
                        "description": "Lista de órdenes obtenida exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Order"
                                            }
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Órdenes obtenidas exitosamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Crear una nueva orden",
                "description": "Crea una orden con los datos del cliente y los productos",
                "operationId": "2123ba58dadb22bd6f2c8081af562230",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "customer_name",
                                    "customer_email",
                                    "items"
                                ],
                                "properties": {
                                    "customer_name": {
                                        "type": "string",
                                        "example": "Juan Pérez"
                                    },
                                    "customer_email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "juan@email.com"
                                    },
                                    "items": {
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "product_name",
                                                "quantity",
                                                "unit_price"
                                            ],
                                            "properties": {
                                                "product_name": {
                                                    "type": "string",
                                                    "example": "Producto A"
                                                },
                                                "quantity": {
                                                    "type": "integer",
                                                    "minimum": 1,
                                                    "example": 2
                                                },
                                                "unit_price": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "minimum": 0,
                                                    "example": 50
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Orden creada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Order"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Orden creada exitosamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Error de validación"
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "customer_name": [
                                                    "El nombre del cliente es obligatorio."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/orders/{id}": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Obtener una orden por ID",
                "description": "Obtiene los detalles de una orden específica con sus items",
                "operationId": "88d27387b1e409e075c18609d4a0c300",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la orden",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Orden obtenida exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Order"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Orden obtenida exitosamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Orden no encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Orden no encontrada"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Order": {
                "title": "Order",
                "description": "Modelo de orden",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "customer_name": {
                        "type": "string",
                        "example": "Juan Pérez"
                    },
                    "customer_email": {
                        "type": "string",
                        "format": "email",
                        "example": "juan@email.com"
                    },
                    "total": {
                        "type": "string",
                        "example": "150.00"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "pending",
                            "processing",
                            "completed",
                            "cancelled"
                        ],
                        "example": "pending"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "items": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/OrderItem"
                        }
                    }
                },
                "type": "object"
            },
            "OrderItem": {
                "title": "OrderItem",
                "description": "Item de una orden",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "order_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "product_name": {
                        "type": "string",
                        "example": "Producto A"
                    },
                    "quantity": {
                        "type": "integer",
                        "example": 2
                    },
                    "unit_price": {
                        "type": "string",
                        "example": "50.00"
                    },
                    "subtotal": {
                        "type": "string",
                        "example": "100.00"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Ingrese el token obtenido del endpoint /api/auth/login",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}